requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
email-validator
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
sqlalchemy-utils
huggingface_hub
requests
